Maybe fix?
Previously the agent would seemingly pathfind correctly, but refuse to park. I cannot test the implementation as I write this, but I think I figured out the problem. The error handling for not finding a spot meant that if the nearest available spot it where the agent already is, the agent would assume it cannot park there, and it would find a new spot. This should now be fixed.